+++ src/deflate/deflate.rs
@@ -1,5 +1,6 @@
 use libz_sys;
 use libc::c_int;
+use std::cmp::max;
 
 /// Decompress a data stream using the DEFLATE algorithm
 pub fn inflate(data: &[u8]) -> Result<Vec<u8>, String> {
@@ -25,10 +26,12 @@ pub fn deflate(data: &[u8], zc: u8, zm: u8, zs: u8, zw: u8) -> Result<Vec<u8>, S
                                                          zw as c_int,
                                                          zm as c_int,
                                                          zs as c_int);
-    let mut output = Vec::with_capacity(data.len() / 20);
+    // Compressed input should be smaller than decompressed, so allocate less than data.len()
+    // However, it needs a minimum capacity in order to handle very small images
+    let mut output = Vec::with_capacity(max(1024, data.len() / 20));
     loop {
         match stream.compress_vec(input.as_mut(), output.as_mut()) {
-            libz_sys::Z_OK => output.reserve(data.len() / 20),
+            libz_sys::Z_OK => output.reserve(max(1024, data.len() / 20)),
             libz_sys::Z_STREAM_END => break,
             c => return Err(format!("Error code on compress: {}", c)),
         }